Updating nodejs version in package.json
Going through this course, trying to deploy this application from a forked repo will always fail because the nodejs version is out of date in comparison to what's being checked by the nodejs buildpack while installing binaries. Thankfully, the errors are detailed, but I believe changing this (or updating the project instructions) would prevent a lot of others from running into the same issue. Thank you!